创建服务特点的增删改查，商品表的生成，创建店铺的经纬度

+++ o2o-basic-parent/o2o-basic-interface/src/main/java/com/jsy/BaseEntity.java
@@ -17,7 +17,7 @@ public class BaseEntity implements Serializable {
     private Long id;
 
     @TableLogic
-    private Long deleted;
+    private int deleted;
 
     @TableField(fill = FieldFill.INSERT)
     @JsonDeserialize(using = LocalDateTimeDeserializer.class)
